+++ src/controllers/__tests__/signoutController.spec.ts
@@ -2,7 +2,6 @@ import { Response } from 'express';
 import { Request } from 'express-jwt';
 import { Cache } from "../../datasources/cache";
 import { revokeAccessToken, revokeRefreshToken } from '../../services/tokenService';
-import { refreshTokenController } from '../refreshTokenController';
 import casual from 'casual';
 import { signoutController } from '../signoutController';
 
@@ -20,7 +19,6 @@ describe('signoutController', () => {
 
     mockRequest = {
       auth: { jti: casual.integer(1, 99999).toString() },
-      headers: { 'authorization': 'Bearer old-access-token' },
     };
     mockResponse = {
       status: jest.fn().mockReturnThis(),
@@ -43,35 +41,32 @@ describe('signoutController', () => {
     await signoutController(mockRequest as Request, mockResponse as Response);
 
     expect(revokeRefreshToken).toHaveBeenCalledWith(mockCache, mockRequest.auth.jti);
-    expect(revokeAccessToken).toHaveBeenCalledWith(mockCache, 'old-access-token');
+    expect(revokeAccessToken).toHaveBeenCalledWith(mockCache, mockRequest.auth.jti);
     expect(mockResponse.clearCookie).toHaveBeenCalledWith('dmspt');
     expect(mockResponse.clearCookie).toHaveBeenCalledWith('dmspr');
     expect(mockResponse.status).toHaveBeenCalledWith(200);
     expect(mockResponse.json).toHaveBeenCalledWith({ success: true, message: 'ok' });
   });
 
-  it('should return 400 if no access token is present', async () => {
+  it('should return 200 if no access token is present', async () => {
     await signoutController(mockRequest as Request, mockResponse as Response);
 
-    expect(mockResponse.status).toHaveBeenCalledWith(400);
-    expect(mockResponse.json)
-      .toHaveBeenCalledWith({ success: false, message: 'Unable to sign out at this time.' });
+    expect(mockResponse.status).toHaveBeenCalledWith(200);
+    expect(mockResponse.json).toHaveBeenCalledWith({ success: true, message: 'ok' });
   });
 
-  it('should return 400 if unable to revoke the refresh token', async () => {
-    mockRequest = { headers: { 'authorization': 'Bearer old-access-token' } };
-
+  it('should return 200 if unable to revoke the refresh token', async () => {
     (revokeRefreshToken as jest.Mock).mockResolvedValue(false);
 
     await signoutController(mockRequest as Request, mockResponse as Response);
 
-    expect(mockResponse.status).toHaveBeenCalledWith(400);
-    expect(mockResponse.json)
-      .toHaveBeenCalledWith({ success: false, message: 'Unable to sign out at this time.' });
+    expect(mockResponse.status).toHaveBeenCalledWith(200);
+    expect(mockResponse.json).toHaveBeenCalledWith({ success: true, message: 'ok' });
   });
 
   it('should return 500 if an unexpected error occurs', async () => {
-    await refreshTokenController(mockRequest as Request, mockResponse as Response);
+    (revokeRefreshToken as jest.Mock).mockImplementation(() => { throw new Error('test error'); });
+    await signoutController(mockRequest as Request, mockResponse as Response);
 
     expect(mockResponse.status).toHaveBeenCalledWith(500);
     expect(mockResponse.json).toHaveBeenCalledWith({ error: 'An unexpected error occurred' });