+++ src/controllers/signoutController.ts
@@ -12,15 +12,14 @@ export const signoutController = async (req: Request, res: Response) => {
       // Delete the refresh token from the cache
       if (await revokeRefreshToken(cache, req.auth.jti)) {
         // Add the access token to the black list so that token is immediately invalidated
-        await revokeAccessToken(cache, req.headers?.authorization?.split(" ")[1]);
+        await revokeAccessToken(cache, req.auth.jti);
 
         // Clear the old cookies from the response
         res.clearCookie('dmspt');
         res.clearCookie('dmspr');
-        res.status(200).json({ success: true, message: 'ok' });
       }
     }
-    res.status(400).json({ success: false, message: 'Unable to sign out at this time.'})
+    res.status(200).json({ success: true, message: 'ok' });
   } catch (err) {
     formatLogMessage(logger).error(err, 'Signout error!');
     res.status(500).json({ error: 'An unexpected error occurred' });